feat: tracking updates and testing
