+++ src/data/services/segment/utils.js
@@ -1,6 +1,6 @@
 /* eslint-disable import/prefer-default-export */
 import { sendTrackEvent } from '@edx/frontend-platform/analytics';
-import { appName } from './constants';
+import { appName } from 'tracking/constants';
 
 export const LINK_TIMEOUT = 300;
