+++ src/tracking/trackers/entitlements.test.js
@@ -32,4 +32,3 @@ describe('entitlements trackers', () => {
     });
   });
 });
-
